% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumudist.R
\name{cumudist}
\alias{cumudist}
\title{calculate proportion less than}
\usage{
cumudist(x)
}
\arguments{
\item{x}{vector of counts}
}
\value{
vector of proportion less than the sieve size.
}
\description{
Note: this function trims the last number for plotting purposes
}
\examples{

}
