% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{describe_kurtosis_fw}
\alias{describe_kurtosis_fw}
\title{Describe Folk and Ward kurtosis}
\usage{
describe_kurtosis_fw(kurtosis)
}
\arguments{
\item{kurtosis}{A numeric vector of Folk and Ward kurtosis values.}
}
\value{
A character vector of kurtosis class labels.
}
\description{
Describe Folk and Ward kurtosis
}
