% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{describe_mean_size_phi}
\alias{describe_mean_size_phi}
\title{Describe mean grain size from phi units}
\usage{
describe_mean_size_phi(phi)
}
\arguments{
\item{phi}{A numeric vector of mean grain sizes in phi units.}
}
\value{
A character vector of grain-size class labels.
}
\description{
Describe mean grain size from phi units
}
