% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{describe_skewness_fw}
\alias{describe_skewness_fw}
\title{Describe Folk and Ward skewness}
\usage{
describe_skewness_fw(skewness)
}
\arguments{
\item{skewness}{A numeric vector of Folk and Ward skewness values.}
}
\value{
A character vector of skewness class labels.
}
\description{
Describe Folk and Ward skewness
}
