% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{describe_sorting_fw}
\alias{describe_sorting_fw}
\title{Describe Folk and Ward sorting}
\usage{
describe_sorting_fw(sorting_phi)
}
\arguments{
\item{sorting_phi}{A numeric vector of Folk and Ward sorting values in phi
units.}
}
\value{
A character vector of sorting class labels.
}
\description{
Describe Folk and Ward sorting
}
