% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elwha-img.R
\docType{data}
\name{elwha_img}
\alias{elwha_img}
\title{Example of grain sizes extracted using grainsizeR}
\format{A data frame with 100 rows and 5 variables:
\describe{
  \item{X}{measurement number}
  \item{measurement_no}{measurement number}
  \item{gsize_cm}{grain size in cm}
  \item{phi}{grain size in phi scale}
  \item{too_small}{binary flag indicating grain size is too small to establish edge while measuring}
  ...
}}
\source{
measured by David Tavernini, imagery from Katie deRego
}
\usage{
elwha_img
}
\description{
A dataset containing 100 grain size samples extracted from Elwha river imagery
}
\keyword{datasets}

