% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elwha-wol.R
\docType{data}
\name{elwha_wol}
\alias{elwha_wol}
\title{Example of grain sizes from Wolman count (sieved)}
\format{A data frame with 100 rows and 5 variables:
\describe{
  \item{Size.class}{sieve size}
  \item{Count}{number of stones sieved}
  ...
}}
\source{
data from Katie deRego
}
\usage{
elwha_wol
}
\description{
A dataset containing 110 grain size samples extracted from Elwha River Wolman pebble count
}
\keyword{datasets}

