% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainsizeR-package.r
\docType{package}
\name{grainsizeR}
\alias{grainsizeR}
\alias{grainsizeR-package}
\title{grainsizeR.}
\description{
Using R and Shiny as a framework for grain size extraction and analysis from imagery.
}
