% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{gs_frac_schemes}
\alias{gs_frac_schemes}
\title{Convenience alias for fraction schemes}
\usage{
gs_frac_schemes()
}
\value{
See \code{\link[=gs_fraction_schemes]{gs_fraction_schemes()}}.
}
\description{
\code{gs_frac_schemes()} is a short convenience alias for
\code{\link[=gs_fraction_schemes]{gs_fraction_schemes()}}. It does not change behavior.
}
