% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gsd_tbl.R
\name{is_gsd_tbl}
\alias{is_gsd_tbl}
\title{Test whether an object is a grain-size distribution tibble}
\usage{
is_gsd_tbl(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{gsd_tbl}, otherwise \code{FALSE}.
}
\description{
Test whether an object is a grain-size distribution tibble
}
