% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{mm_to_phi}
\alias{mm_to_phi}
\title{Convert millimeters to phi units}
\usage{
mm_to_phi(x)
}
\arguments{
\item{x}{A numeric vector of sizes in millimeters.}
}
\value{
A numeric vector of phi sizes.
}
\description{
Phi is calculated as \code{-log2(size_mm)}.
}
