% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{mm_to_um}
\alias{mm_to_um}
\title{Convert millimeters to micrometers}
\usage{
mm_to_um(x)
}
\arguments{
\item{x}{A numeric vector of sizes in millimeters.}
}
\value{
A numeric vector of sizes in micrometers.
}
\description{
Convert millimeters to micrometers
}
