% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{phi_to_mm}
\alias{phi_to_mm}
\title{Convert phi units to millimeters}
\usage{
phi_to_mm(x)
}
\arguments{
\item{x}{A numeric vector of phi sizes.}
}
\value{
A numeric vector of sizes in millimeters.
}
\description{
Convert phi units to millimeters
}
