% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{phi_to_um}
\alias{phi_to_um}
\title{Convert phi units to micrometers}
\usage{
phi_to_um(x)
}
\arguments{
\item{x}{A numeric vector of phi sizes.}
}
\value{
A numeric vector of sizes in micrometers.
}
\description{
Convert phi units to micrometers
}
