% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_grainsizeR.R
\name{run.grainsizeR}
\alias{run.grainsizeR}
\title{Grain size extaction from photographs}
\usage{
run.grainsizeR()
}
\value{
dataframe of grain sizes
}
\description{
Initialize shiny app for the grain size extraction from photographs
}
\examples{
#' run.grainsizeR()  # Open app to begin grain size extraction interface
}

