% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sieve_mm.R
\name{sieve.mm}
\alias{sieve.mm}
\title{Sieve data (mm)}
\usage{
sieve.mm(x, sieve_sizes = c(8, 11, 16, 22.6, 32, 45, 64, 90, 128, 180, 256,
  360, 512, 1024))
}
\arguments{
\item{x}{a list of grain size measurements}

\item{sieve_sizes}{a vector containing desired sieve sizes in order of ascending size}
}
\value{
a dataframe of sieve sizes and associated counts of particles sieved at that size
}
\description{
Sieve data (mm)
}
\examples{

}
