% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sieve_phi.R
\name{sieve.phi}
\alias{sieve.phi}
\title{sieve data (phi)}
\usage{
sieve.phi(x, sieve_sizes = -log2(c(8, 11, 16, 22.6, 32, 45, 64, 90, 128, 180,
  256, 360, 512, 1024)))
}
\arguments{
\item{x}{vector containing individual measurements}

\item{sieve_sizes}{vector containing sieve sizes in phi scale in ascending order}
}
\value{
dataframe containing sieve sizes and according number of particles retained within each sieve
}
\description{
sieve data (phi)
}
\examples{

}

