% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_dx.R
\name{stat.dx}
\alias{stat.dx}
\title{percentile return}
\usage{
stat.dx(sieve_size, cumu, d = 0.5, phi = FALSE)
}
\arguments{
\item{sieve_size}{vector of sieve sizes}

\item{cumu}{vector of corresponding cumulative distribution values}

\item{d}{percentile value (decimal)}

\item{phi}{logical to state if using phi units or not}
}
\value{
value of grain size of desired percentile value
}
\description{
percentile return
}
\examples{

}

