% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_kurt.R
\name{stat.kurt}
\alias{stat.kurt}
\title{Measure kurtosis from cumulative distribution data}
\source{
see Bunte and Abt (2001) for examples and equations. Bunte, Kristin, and Steven R. Abt. "Sampling surface and subsurface particle-size distributions in wadable gravel-and cobble-bed streams for analyses in sediment transport, hydraulics, and streambed monitoring." (2001).
}
\usage{
stat.kurt(sieve_size, cumu, units = "phi", method = "all")
}
\arguments{
\item{sieve_size}{vector of sieve sizes related to input cumulative distribution}

\item{cumu}{cumulative distribution pertaining to sieve sizes}

\item{units}{sieve size units "phi" = phi scale, "mm" = milimeter scale}

\item{method}{Method to calculate kurtosis. "fw" = Folk and Ward, 1957. This method is typically used for phi units and is fairly standard; "inm" = Inman, 1952. For phi scale and focuses on tails of distribution; "tra" = Trask, 1932. For metric (mm) data.; "sqrt" = Square root method for geometric approach in metric units;"log" = log transform metric data. Geometric approach. "all" = compute all methods into a data frame.}
}
\value{
a numeric value or dataframe according to method chosen
}
\description{
Measure kurtosis from cumulative distribution data
}
\examples{

}
