% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mean.R
\name{stat.mean}
\alias{stat.mean}
\title{Compute mean grain size from cumulative size distribution}
\source{
see Bunte and Abt (2001) for examples and equations. Bunte, Kristin, and Steven R. Abt. "Sampling surface and subsurface particle-size distributions in wadable gravel-and cobble-bed streams for analyses in sediment transport, hydraulics, and streambed monitoring." (2001).
}
\usage{
stat.mean(sieve_size, cumu, units = "phi", method = "all")
}
\arguments{
\item{sieve_size}{vector of sieve sizes related to input cumulative distribution}

\item{cumu}{cumulative distribution pertaining to sieve sizes}

\item{units}{sieve size units "phi" = phi scale, "mm" = milimeter (metric) scale}

\item{method}{Algebraic: "inm" = Inman (1952) very simple calculation focus on central data;"tra" = Trask (1932) use interquartile values;"fw" = Folk and Ward (1957) incorporate median value to Inman equation;"brig" = Briggs (1977) use all percentile values (10, 20, 30...80, 90) to incorprate distribution;"mccam" = McCammon (1962) Similar to Briggs but on the 5th percentiles. Geometric: "sqrt" = Geometric approach similar to Inman but with metric log-normal distributions;"cubr" = Similar to square root but incorporate D50;"log" = Log transform metric data to calculate and backtransform result to original units (works for log-normal distributions); "all" = All outputs into a dataframe}
}
\value{
dataframe or value containing specified measures of mean
}
\description{
Compute mean grain size from cumulative size distribution
}
\examples{

}

