% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_sorting.R
\name{stat.sorting}
\alias{stat.sorting}
\title{Compute sorting coefficient from cumulative distribution data}
\source{
see Bunte and Abt (2001) for examples and equations. Bunte, Kristin, and Steven R. Abt. "Sampling surface and subsurface particle-size distributions in wadable gravel-and cobble-bed streams for analyses in sediment transport, hydraulics, and streambed monitoring." (2001).
}
\usage{
stat.sorting(sieve_size, cumu, units = "phi", method = "all")
}
\arguments{
\item{sieve_size}{vector of sieve sizes related to input cumulative distribution}

\item{cumu}{cumulative distribution pertaining to sieve sizes}

\item{units}{sieve size units "phi" = phi scale, "mm" = milimeter (metric) scale}

\item{method}{Method to calculate sorting coefficient. Only algebraic methods available. "inm" = Inman, 1952. Algebraic method using central values; "fw" = Folk and Ward, 1957. Algebraic method to account for tail values. "all" = dataframe output with above calculations.}
}
\value{
a value or dataframe containing specified sorting coefficients.
}
\description{
Compute sorting coefficient from cumulative distribution data
}
\examples{

}

