% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{um_to_mm}
\alias{um_to_mm}
\title{Convert micrometers to millimeters}
\usage{
um_to_mm(x)
}
\arguments{
\item{x}{A numeric vector of sizes in micrometers.}
}
\value{
A numeric vector of sizes in millimeters.
}
\description{
Convert micrometers to millimeters
}
