% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{um_to_phi}
\alias{um_to_phi}
\title{Convert micrometers to phi units}
\usage{
um_to_phi(x)
}
\arguments{
\item{x}{A numeric vector of sizes in micrometers.}
}
\value{
A numeric vector of phi sizes.
}
\description{
Convert micrometers to phi units
}
